--- a/packages/vitest/src/runtime/execute.ts
+++ b/packages/vitest/src/runtime/execute.ts
@@ -1,6 +1,6 @@
 import vm from 'node:vm'
 import { pathToFileURL } from 'node:url'
-import { readFile } from 'node:fs/promises'
+import { readFileSync } from 'node:fs'
 import type { ModuleCacheMap } from 'vite-node/client'
 import { DEFAULT_REQUEST_STUBS, ViteNodeRunner } from 'vite-node/client'
 import { isInternalRequest, isNodeBuiltin, isPrimitive, toFilePath } from 'vite-node/utils'
@@ -107,7 +107,7 @@ export async function startVitestExecutor(options: ContextExecutorOptions) {
 
       const result = await rpc().fetch(id, getTransformMode())
       if (result.id && !result.externalize) {
-        const code = await readFile(result.id, 'utf-8')
+        const code = readFileSync(result.id, 'utf-8')
         return { code }
       }
       return result