--- a/packages/vitest/src/utils/base.ts
+++ b/packages/vitest/src/utils/base.ts
@@ -169,3 +169,14 @@ export function escapeRegExp(s: string) {
 export function wildcardPatternToRegExp(pattern: string): RegExp {
   return new RegExp(`^${pattern.split('*').map(escapeRegExp).join('.*')}$`, 'i')
 }
+
+// port from nanoid
+// https://github.com/ai/nanoid
+const urlAlphabet
+  = 'useandom-26T198340PX75pxJACKVERYMINDBUSHWOLF_GQZbfghjklqvwyzrict'
+export function nanoid(size = 21) {
+  let id = ''
+  let i = size
+  while (i--) id += urlAlphabet[(Math.random() * 64) | 0]
+  return id
+}