### Description

This PR makes forks pool 2 times faster on https://github.com/vuejs/core monorepo (using M1 Air):

Before:

```
threads:
Duration  16.17s (transform 5.28s, setup 740ms, collect 54.54s, tests 7.42s, environment 14.06s, prepare 13.76s)

forks:
Duration  28.68s (transform 6.17s, setup 1.30s, collect 138.67s, tests 5.53s, environment 11.92s, prepare 11.67s)
```

After:

```
threads:
Duration  16.29s (transform 6.79s, setup 1.81s, collect 55.58s, tests 6.79s, environment 12.41s, prepare 14.32s)

forks:
Duration  18.89s (transform 7.70s, setup 811ms, collect 65.05s, tests 7.40s, environment 14.14s, prepare 14.98s)
```

<!-- Please insert your description here and provide especially info about the "what" this PR is solving -->

<!-- You can also add additional context here -->

### Please don't delete this checklist! Before submitting the PR, please make sure you do the following:
- [ ] It's really useful if your PR references an issue where it is discussed ahead of time. If the feature is substantial or introduces breaking changes without a discussion, PR might be closed.
- [ ] Ideally, include a test that fails without this PR but passes with it.
- [ ] Please, don't make changes to `pnpm-lock.yaml` unless you introduce a new test example.

### Tests
- [ ] Run the tests with `pnpm test:ci`.

### Documentation
- [ ] If you introduce new functionality, document it. You can run documentation with `pnpm run docs` command.

### Changesets
- [ ] Changes in changelog are generated from PR name. Please, make sure that it explains your changes in an understandable manner. Please, prefix changeset messages with `feat:`, `fix:`, `perf:`, `docs:`, or `chore:`.