@sheremet-va 2024-04-23
Changed `readFileSync` to `promises.readFile` - need to make new benchmarks.