@AriPerkkio 2024-04-23
> Changed `readFileSync` to `promises.readFile` - need to make new benchmarks now.

I tested that yesterday and saw worse perf - no idea why 🙃