@sheremet-va 2024-04-23
> > Changed `readFileSync` to `promises.readFile` - need to make new benchmarks now.
> 
> I tested that yesterday and saw worse perf - no idea why 🙃

Probably because there are not a lot of files loading in parallel so nothing is really blocked most of the time. I will revert it to a sync one for now.