@sheremet-va 2024-04-23
By the way, `--coverage` is also quite slower on forks:

```diff
- Duration  32.95s (transform 24.10s, setup 1.50s, collect 154.66s, tests 8.55s, environment 11.48s, prepare 16.06s)
+ Duration  44.26s (transform 22.62s, setup 1.51s, collect 163.91s, tests 8.04s, environment 10.14s, prepare 17.95s)
```

Looks like the communication part (`onAfterSuiteRun`) is not displayed at all