@sheremet-va 2024-04-23
> @AriPerkkio looks like the coverage is still broken 🤔

Ok, I was also removing `code` which broke v8 coverage. Should be fine now.