@sheremet-va 2024-04-23
> The coverage results are sent to the main thread via rpc.

But how big are the messages?