@AriPerkkio 2024-04-23
> > The coverage results are sent to the main thread via rpc.
> 
> But how big are the messages?


Comment out these parts and check `coverage/.tmp`:

https://github.com/vitest-dev/vitest/blob/23f29cea6491f3ecde8b885413f4c46989ab3bfc/packages/coverage-v8/src/provider.ts#L247-L248

On Vue repo:
- v8: 177 files, total 43M
- istanbul: 177 files, 362M