@sheremet-va 2024-04-24
> On Vue repo

Yeah, that's a lot of data to code/decode all the time